+++ example-bareflask/rootless-start.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -e
 
-[ -f .env ] && export $(grep -v '^#' .env | xargs)
+[ -f .env ] && source .env
 if [ "$rootless" -ne 1 ] 2>/dev/null; then
   echo "Exiting because \$rootless is not set to 1."
   exit 1
@@ -10,42 +10,59 @@ if [ -z "$host_data_dir" ]; then
   exit 2
 fi
 
-ROOTLESS_NAME="$challenge_docker_name-rootless-daemon"
-ROOTLESS_NETWORK="$challenge_docker_name-rootless-network"
-DIND_IMAGE="docker:25.0-dind-rootless"
-RUN_DIR="$host_data_dir/rootless_user/"
-DOCKER_DIR="$host_data_dir/rootless_docker/"
+RUN_DIR="$PWD/run"
+VOL_DIR="${host_data_dir}/vol"
 
-sudo rm -rf "$RUN_DIR"
-sudo mkdir -p "$RUN_DIR"
-sudo chown 1000:1000 "$RUN_DIR"
-sudo chmod 700 "$RUN_DIR"
-sudo mkdir -p "$DOCKER_DIR"
-sudo chown 1000:1000 "$DOCKER_DIR"
-sudo chmod 755 "$DOCKER_DIR"
+start() {
+	sudo mkdir -p "${VOL_DIR}"
+	sudo chown 1000:1000 "${VOL_DIR}"
+	sudo chmod 755 "${VOL_DIR}"
 
-sudo docker rm -f "$ROOTLESS_NAME"
-sudo docker network create "$ROOTLESS_NETWORK" || true
-sudo docker run -d -v "$RUN_DIR":/run/user/1000 -v "$DOCKER_DIR":/home/rootless/.local/share/docker \
-  --name "$ROOTLESS_NAME" --network "$challenge_docker_name-rootless-network" \
-  --restart=always -p "127.0.0.1:$port:$port" --privileged "$DIND_IMAGE"
+	docker compose -f docker-compose-rootless.yml up -d --build --remove-orphans
+	
+	set +e
+	for (( i=1; i<=30; i++ ))
+	do
+	  HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" --unix-socket "$RUN_DIR"/docker.sock http://localhost/_ping)
+	  if [ "$HTTP_STATUS" -eq 200 ]; then
+	    echo "OK!"
+	    break
+	  else
+	    echo "Failed (HTTP $HTTP_STATUS)... ($i/30)"
+	  fi
+	  sleep 1
+	done
+	set -e
 
-echo "Waiting for rootless daemon to be alive"
+	export DOCKER_HOST=unix://"$RUN_DIR"/docker.sock
+	sudo "DOCKER_HOST=$DOCKER_HOST" docker compose up -d --build
+	echo "Set DOCKER_HOST=$DOCKER_HOST when operating inside rootless container..."
+}
 
-set +e
-for (( i=1; i<=30; i++ ))
-do
-  HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" --unix-socket "$RUN_DIR"/docker.sock http://localhost/_ping)
-  if [ "$HTTP_STATUS" -eq 200 ]; then
-    echo "OK!"
-    break
-  else
-    echo "Failed (HTTP $HTTP_STATUS)... ($i/30)"
-  fi
-  sleep 1
-done
-set -e
+stop() {
+	DOCKER_SOCK="$RUN_DIR"/docker.sock
+	DOCKER_HOST=unix://"$RUN_DIR"/docker.sock
+	if socat -u OPEN:/dev/null UNIX-CONNECT:"$DOCKER_SOCK"; then
+		sudo "DOCKER_HOST=$DOCKER_HOST" docker compose down
+	fi
+	sudo docker compose -f docker-compose-rootless.yml down
+}
 
-export DOCKER_HOST=unix://"$RUN_DIR"/docker.sock
-sudo "DOCKER_HOST=$DOCKER_HOST" docker compose up -d
-echo "Set DOCKER_HOST=$DOCKER_HOST when operating inside rootless container..."
+update() {
+	DOCKER_HOST=unix://"$RUN_DIR"/docker.sock
+	DOCKER_SOCK="$RUN_DIR"/docker.sock
+	if socat -u OPEN:/dev/null UNIX-CONNECT:"$DOCKER_SOCK"; then
+		sudo "DOCKER_HOST=$DOCKER_HOST" docker compose up -d --build --remove-orphans
+	else
+		start
+	fi
+}
+
+case "${1}" in
+	start|stop|update)
+		"${1}"
+		;;
+	*)
+		echo "Usage: start|stop|update"
+		;;
+esac